Created pkg/common/crypto.
