+++ pkg/common/crypto/testdata/clientkey.pem
